:recycle: question write 리팩터링
