--- src/components/common/ToastEditor.tsx
@@ -1,30 +1,0 @@
-import { Editor } from '@toast-ui/react-editor';
-import '@toast-ui/editor/dist/toastui-editor.css';
-import { uploadImage } from '@/api/images';
-
-export default function ToastEditor() {
-  const addImageBlobHook = async (file, callback) => {
-    const { data } = await uploadImage(file);
-    callback(data.url, '이미지');
-  };
-
-  return (
-    <Editor
-      placeholder="내용을 입력해주세요."
-      previewStyle="vertical"
-      initialEditType="wysiwyg"
-      height="350px"
-      toolbarItems={[
-        // 툴바 옵션 설정
-        ['heading', 'bold', 'italic', 'strike'],
-        ['hr', 'quote'],
-        ['ul', 'ol', 'task', 'indent', 'outdent'],
-        ['table', 'image', 'link'],
-        ['code', 'codeblock'],
-      ]}
-      hooks={{
-        addImageBlobHook,
-      }}
-    />
-  );
-}
